Fixed error in index.html and edited styles.css
Anchor tags in header_links div did not surround the text and would
cause an issue when trying to style it with styles.css. Styles were
added to the content of section_1, which was also given a display of
flex. Lists items were also edited to remove the bullet points and
default styles.