added properties to styles.css
The section_3 class and the elements nested within it were updated to
include the display type of flex, add padding to the section, and
make other changes to enhance the presentation.The div was also centered
to keep content in the middle of that container. A width was also added
to the container to prevent the text from trailing too long.