Updated and added new styles to styles.css
The section_4 div was given more padding, padding and a width was given
to the s4_container. This was done so that the text and button could be
spaced evenly. the text inside the div was also given a color and a
margin of 0 so that the header and paragraph content will be closer
together.